src/api/user/infrastructure/http/dtos/verify_account/pydantic_verify_account_response_dto.py
from pydantic import BaseModel

from src.api.user.infrastructure.persistence.models.sqlmodel_user_model import (
    SqlModelUserModel,
)


class PydanticVerifyAccountResponseDTO(BaseModel):
    user: SqlModelUserModel
    session_token: str
