pyproject.toml
[tool.poetry]
name = "synclife-api"
version = "0.0.1"
description = "FastAPI project started"
authors = ["Giuliano Poeta <giuliano.poeta.laboral@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.1.1"
black = "^24.10.0"
fastapi = { extras = ["standard"], version = "^0.115.4" }
bcrypt = "^4.2.0"
sqlmodel = "^0.0.22"
psycopg2 = "^2.9.10"
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
zxcvbn = "^4.4.28"
isort = "^5.13.2"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
types-zxcvbn = "^4.4.1.20240106"

[tool.mypy]
allow_redefinition = true         # Permite redefinir métodos (útil para @property y @setter)
disallow_untyped_calls = true     # No permite llamadas a funciones sin anotaciones
disallow_untyped_defs = true      # Obliga a tipar todas las funciones
disallow_incomplete_defs = true   # Obliga a definir completamente argumentos y retornos
no_implicit_optional = true       # No asume Optional implícitamente
strict_optional = true            # Maneja correctamente los valores None
ignore_missing_imports = false    # Asegura que todos los imports estén tipados

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"                                                              # Compatible con Black
line_length = 88                                                               # Mismo valor que en Black
multi_line_output = 3
include_trailing_comma = true
known_first_party = ["synclife_api"]                                           # Ajustá el nombre según tu módulo
known_third_party = ["fastapi", "sqlalchemy", "bcrypt", "psycopg2", "alembic"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
